+++ docs/Changelog.md
@@ -11,6 +11,7 @@
  * Server `getConnections()`, `getReadableConnections()`, `getWritableConnections()` (@sirn-se)
  * `onHandshake(...)` listener (will deprecate `onConnect(...)`) (@sirn-se)
  * Server `setMaxConnections(int|null)` (@sirn-se)
+ * Server `shutdown()` orderly close server (@sirn-se)
 
 ### `3.0.0`